fix(bridge): Pretty-print request errors (#7652)
